Set the always-use-proxy to true if a proxy is selected
We now assume the proxy is by default a full proxy even if it is the Tor gate socks proxy,
We declare --proxy an early option and disallow the proxy address itself to be a FQDN.
So that we only overide the false value when we need to have setups without proxy.
The user can override this if he sets --always-use-proxy=false after
the --proxy cmd to route only tor onion address traffic over the proxy.

Changelog-Changed: Set the --always-use-proxy option to true if a proxy is selected.